Suppress PVS Studio false-positive warning for dr_wav
Basis for why this is being suppressed, per @mackron (12 Nov 2020):

---
This check is specifically for 32-bit builds where that should certainly be
possible. Won't be possible on 64-bit, but I don't want to be doing a
compile-time check on that to avoid the warning.

I made a tiny change to this here which almost certainly won't fix this
message. I've decided that I'm leaving the code as-is because I can't think of
a clean way to silence it.
---

Because this check is valid in some cases, we consider PVS's
assessment incorrect and thus a false-positive.

Signed-off-by: kcgen <1557255+kcgen@users.noreply.github.com>